Perfect hiring problem fixed.

+++ src/main/java/net/burakince/hackerrank/perfect_hiring/Solution.java
@@ -7,18 +7,18 @@ public class Solution {
 	public static void main(String[] args) {
 		Scanner in = new Scanner(System.in);
 
-		short candidates = in.nextShort();
-		int patience = in.nextInt();
+		int candidates = in.nextInt();
+		long patience = in.nextLong();
 		byte losePatience = in.nextByte();
 		long decide = 0;
-		short decideId = 0;
+		int decideId = 0;
 
-		for (short n = 0; n < candidates; n++) {
+		for (int n = 0; n < candidates; n++) {
 			int score = in.nextInt();
 			long newDecide = score * patience;
 			if (newDecide > decide) {
 				decide = newDecide;
-				decideId = (short) (n + 1);
+				decideId = n + 1;
 			}
 			patience -= losePatience;
 		}
